Fix the tests on 1.3 and HEAD
In the latest Django master code, RequestFactory.put behaves fundamentally differently than it did pre-1.5.  By default, it expects an octet string as opposed to a dictionary that it will encode like a multipart form.  So, for 1.5 and on, we have to be explicit about the multipart type and pre-encode the data.  However, pre-1.5 Django expects a dictionary if the content type is multipart.  So, the cleanest thing to do is explicitly handle the versions independently.